ref(api): Errors submodule
This commit adds a new submodule to the API module, `errors`, which contains error types. Future PRs will split the `errors` module into submodules for each error type.
